lib: deprecate _http_*
