+++ benchmark/http/check_invalid_header_char.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const common = require('../common.js');
-const _checkInvalidHeaderChar = require('_http_common')._checkInvalidHeaderChar;
 
 const groupedInputs = {
   // Representative set of inputs from an AcmeAir benchmark run:
@@ -51,9 +50,13 @@ const inputs = [
 const bench = common.createBenchmark(main, {
   input: inputs.concat(Object.keys(groupedInputs)),
   n: [1e6],
+}, {
+  flags: ['--expose-internals', '--no-warnings'],
 });
 
 function main({ n, input }) {
+  const checkInvalidHeaderChar = require('internal/http/common')._checkInvalidHeaderChar;
+
   let inputs = [input];
   if (Object.hasOwn(groupedInputs, input)) {
     inputs = groupedInputs[input];
@@ -62,7 +65,7 @@ function main({ n, input }) {
   const len = inputs.length;
   bench.start();
   for (let i = 0; i < n; i++) {
-    _checkInvalidHeaderChar(inputs[i % len]);
+    checkInvalidHeaderChar(inputs[i % len]);
   }
   bench.end(n);
 }
