+++ benchmark/http/check_is_http_token.js
@@ -1,7 +1,6 @@
 'use strict';
 
 const common = require('../common.js');
-const _checkIsHttpToken = require('_http_common')._checkIsHttpToken;
 
 const bench = common.createBenchmark(main, {
   key: [
@@ -38,12 +37,16 @@ const bench = common.createBenchmark(main, {
     'alternate-protocol:', // slow bailout
   ],
   n: [1e6],
+}, {
+  flags: ['--expose-internals', '--no-warnings'],
 });
 
 function main({ n, key }) {
+  const checkIsHttpToken = require('internal/http/common')._checkIsHttpToken;
+
   bench.start();
   for (let i = 0; i < n; i++) {
-    _checkIsHttpToken(key);
+    checkIsHttpToken(key);
   }
   bench.end(n);
 }
