+++ lib/http.js
@@ -26,22 +26,22 @@ const {
 } = primordials;
 
 const { validateInteger, validateObject } = require('internal/validators');
-const httpAgent = require('_http_agent');
-const { ClientRequest } = require('_http_client');
-const { methods, parsers } = require('_http_common');
-const { IncomingMessage } = require('_http_incoming');
+const httpAgent = require('internal/http/agent');
+const { ClientRequest } = require('internal/http/client');
+const { methods, parsers } = require('internal/http/common');
+const { IncomingMessage } = require('internal/http/incoming');
 const { ERR_PROXY_INVALID_CONFIG } = require('internal/errors').codes;
 const {
   validateHeaderName,
   validateHeaderValue,
   OutgoingMessage,
-} = require('_http_outgoing');
+} = require('internal/http/outgoing');
 const {
   _connectionListener,
   STATUS_CODES,
   Server,
   ServerResponse,
-} = require('_http_server');
+} = require('internal/http/server');
 const {
   parseProxyUrl,
   getGlobalAgent,
