+++ lib/https.js
@@ -53,15 +53,15 @@ const {
   kWaitForProxyTunnel,
   getGlobalAgent,
 } = require('internal/http');
-const { Agent: HttpAgent } = require('_http_agent');
+const { Agent: HttpAgent } = require('internal/http/agent');
 const {
   httpServerPreClose,
   Server: HttpServer,
   setupConnectionsTracking,
   storeHTTPOptions,
   _connectionListener,
-} = require('_http_server');
-const { ClientRequest } = require('_http_client');
+} = require('internal/http/server');
+const { ClientRequest } = require('internal/http/client');
 let debug = require('internal/util/debuglog').debuglog('https', (fn) => {
   debug = fn;
 });
