+++ lib/internal/child_process.js
@@ -148,9 +148,9 @@ const handleConversion = {
         socket.setTimeout(0);
 
         if (freeParser === undefined)
-          freeParser = require('_http_common').freeParser;
+          freeParser = require('internal/http/common').freeParser;
         if (HTTPParser === undefined)
-          HTTPParser = require('_http_common').HTTPParser;
+          HTTPParser = require('internal/http/common').HTTPParser;
 
         // In case of an HTTP connection socket, release the associated
         // resources
