+++ lib/internal/http2/compat.js
@@ -58,7 +58,8 @@ const {
   assertValidPseudoHeader,
   getAuthority,
 } = require('internal/http2/util');
-const { _checkIsHttpToken: checkIsHttpToken } = require('_http_common');
+
+const { _checkIsHttpToken: checkIsHttpToken } = require('internal/http/common');
 
 const kBeginSend = Symbol('begin-send');
 const kState = Symbol('state');
