+++ lib/internal/http2/core.js
@@ -48,8 +48,14 @@ const { Duplex } = require('stream');
 const tls = require('tls');
 const { setImmediate, setTimeout, clearTimeout } = require('timers');
 
-const { kIncomingMessage } = require('_http_common');
-const { kServerResponse, Server: HttpServer, httpServerPreClose, setupConnectionsTracking } = require('_http_server');
+const { kIncomingMessage } = require('internal/http/common');
+const {
+  kServerResponse,
+  Server: HttpServer,
+  httpServerPreClose,
+  setupConnectionsTracking,
+} = require('internal/http/server');
+
 const JSStreamSocket = require('internal/js_stream_socket');
 
 const {
