+++ lib/internal/http2/util.js
@@ -16,7 +16,7 @@ const {
 
 const {
   _checkIsHttpToken: checkIsHttpToken,
-} = require('_http_common');
+} = require('internal/http/common');
 
 const binding = internalBinding('http2');
 const {
