+++ test/async-hooks/test-httpparser.request.js
@@ -1,3 +1,5 @@
+// Flags: --expose-internals --no-warnings
+
 'use strict';
 
 const common = require('../common');
@@ -9,7 +11,7 @@ const { checkInvocations } = require('./hook-checks');
 const hooks = initHooks();
 hooks.enable();
 
-const { HTTPParser } = require('_http_common');
+const { HTTPParser } = require('internal/http/common');
 
 const REQUEST = HTTPParser.REQUEST;
