+++ test/parallel/test-http-invalidheaderfield2.js
@@ -1,8 +1,10 @@
+// Flags: --expose-internals --no-warnings
+
 'use strict';
 require('../common');
 const assert = require('assert');
 const inspect = require('util').inspect;
-const { _checkIsHttpToken, _checkInvalidHeaderChar } = require('_http_common');
+const { _checkIsHttpToken, _checkInvalidHeaderChar } = require('internal/http/common');
 
 // Good header field names
 [
