+++ test/parallel/test-http-parser-memory-retention.js
@@ -1,9 +1,10 @@
+// Flags: --expose-internals --no-warnings
 'use strict';
 
 const common = require('../common');
 const assert = require('assert');
 const http = require('http');
-const { HTTPParser } = require('_http_common');
+const { HTTPParser } = require('internal/http/common');
 
 // Test that the `HTTPParser` instance is cleaned up before being returned to
 // the pool to avoid memory retention issues.
