+++ test/parallel/test-https-server-async-dispose.js
@@ -1,3 +1,4 @@
+// Flags: --expose-internals --no-warnings
 'use strict';
 
 const common = require('../common');
@@ -7,7 +8,7 @@ if (!common.hasCrypto)
 
 const assert = require('assert');
 const { createServer } = require('https');
-const { kConnectionsCheckingInterval } = require('_http_server');
+const { kConnectionsCheckingInterval } = require('internal/http/server');
 
 const server = createServer();
